requirement.txt
google-generativeai>=0.3.2   # Gemini Flash 1.5 supported
praw>=7.7.1                  # Reddit API wrapper
